Move tests to a correct file
